refactor: merge flux and libstd into one exportable crate "flux" (#2676)
* refactor: merge flux and libstd into one exportable crate "flux"

1. Renamed flux to core.
2. Rename libstd to flux.
3. From new flux: pub use core::*, pub use core::ast, etc;
4. Fixed up the Cargo.toml files.
5. Fixed up the use statements needing fixing.
6. Fixed up the Makefile.
7. Moved public C API components in the core crate to the flux crate

refs #2655
