+++ libflux/src/core/Cargo.toml
@@ -1,0 +1,36 @@
+[package]
+name = "core"
+version = "0.4.0"
+authors = ["Nathaniel Cook <nathaniel@influxdata.com>"]
+edition = "2018"
+
+[lib]
+name = "core"
+path = "lib.rs"
+crate-type = ["rlib"]
+
+[features]
+default = ["strict"]
+
+strict = []
+
+[dependencies]
+serde = "^1.0.59"
+serde_derive = "^1.0.59"
+serde_json = "1.0"
+serde-aux = "0.6.1"
+wasm-bindgen = { version = "0.2.55", features = ["serde-serialize"] }
+chrono = { version = "0.4", features = ["serde"] }
+regex = "1"
+maplit = "1.0.2"
+flatbuffers = "0.6.1"
+derivative = "1.0.3"
+walkdir = "2.2.9"
+
+[dev-dependencies]
+colored = "1.8"
+pretty_assertions = "0.6.1"
+
+[build-dependencies]
+cc = "1.0"
+bindgen = "0.49.0"
