+++ libflux/src/core/formatter/tests.rs
