+++ libflux/src/core/scanner/mod.rs
