+++ libflux/src/core/scanner/scanner.h
