+++ libflux/src/core/scanner/scanner.rl
