+++ libflux/src/core/scanner/tests.rs
