+++ libflux/src/core/scanner/unicode.rl.COPYING
