--- libflux/src/libstd/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "libstd"
-version = "0.1.0"
-authors = ["jlapacik <joshua.lapacik@pacbell.net>"]
-edition = "2018"
-
-[lib]
-name = "libstd"
-path = "lib.rs"
-crate-type = ["rlib", "staticlib", "cdylib"]
-
-[dependencies]
-flux = { path = "../flux" }
-flatbuffers = "0.6.0"
-
-[dev-dependencies]
-maplit = "1.0.2"
-
-[build-dependencies]
-flux = { path = "../flux" }
-flatbuffers = "0.6.0"
